+++ packages/core/src/util/token.ts
@@ -26,12 +26,12 @@ import { jwtVerify, createRemoteJWKSet } from "jose";
 type WithStack = { stack: string };
 
 /**
- * Extract a WebID from an ID token payload based on https://github.com/solid/webid-oidc-spec.
+ * Extract a WebID and the clientID from an ID token payload based on https://github.com/solid/webid-oidc-spec.
  * Note that this does not yet implement the user endpoint lookup, and only checks
- * for `webid` or IRI-like `sub` claims.
+ * for `webid`, `azp` or IRI-like `sub` claims.
  *
  * @param idToken the payload of the ID token from which the WebID can be extracted.
- * @returns a WebID extracted from the ID token.
+ * @returns an object with entries webId and clientId extracted from the ID token.
  * @internal
  */
 export async function getWebidFromTokenPayload(
@@ -39,8 +39,9 @@ export async function getWebidFromTokenPayload(
   jwksIri: string,
   issuerIri: string,
   clientId: string,
-): Promise<string> {
+): Promise<{ webId: string; clientId?: string }> {
   let payload: JWTPayload;
+  let clientIdInPayload: string | undefined;
   try {
     const { payload: verifiedPayload } = await jwtVerify(
       idToken,
@@ -55,8 +56,14 @@ export async function getWebidFromTokenPayload(
     throw new Error(`Token verification failed: ${(e as WithStack).stack}`);
   }
 
+  if (typeof payload.azp === "string") {
+    clientIdInPayload = payload.azp;
+  }
   if (typeof payload.webid === "string") {
-    return payload.webid;
+    return {
+      webId: payload.webid,
+      clientId: clientIdInPayload,
+    };
   }
   if (typeof payload.sub !== "string") {
     throw new Error(
@@ -71,7 +78,10 @@ export async function getWebidFromTokenPayload(
     // as specified by the Identity Provider.
     // eslint-disable-next-line no-new
     new URL(payload.sub);
-    return payload.sub;
+    return {
+      webId: payload.sub,
+      clientId: clientIdInPayload,
+    };
   } catch (e) {
     throw new Error(
       `The token has no 'webid' claim, and its 'sub' claim of [${payload.sub}] is invalid as a URL - error [${e}].`,