+++ packages/node/src/login/oidc/oidcHandlers/ClientCredentialsOidcHandler.spec.ts
@@ -145,11 +145,14 @@ const setupOidcClientMock = (tokenSet: OpenidClient.TokenSet) => {
   Issuer.mockReturnValueOnce(mockedIssuer);
 };
 
-const setupGetWebidMock = (webid: string) => {
+const setupGetWebidMock = (webid: string, clientid?: string) => {
   const { getWebidFromTokenPayload } = jest.requireMock(
     "@inrupt/solid-client-authn-core",
   ) as jest.Mocked<typeof SolidClientAuthnCore>;
-  getWebidFromTokenPayload.mockResolvedValueOnce(webid);
+  getWebidFromTokenPayload.mockResolvedValueOnce({
+    webId: webid,
+    clientId: clientid,
+  });
 };
 
 describe("ClientCredentialsOidcHandler", () => {
@@ -319,6 +322,7 @@ describe("handle", () => {
   it("builds a fetch authenticated with a DPoP token if appropriate", async () => {
     const tokens = mockDpopTokens();
     setupOidcClientMock(tokens);
+    setupGetWebidMock("https://my.webid/", "some client ID");
     const clientCredentialsOidcHandler = new ClientCredentialsOidcHandler(
       mockDefaultTokenRefresher(),
     );
@@ -347,6 +351,7 @@ describe("handle", () => {
   it("builds a fetch authenticated with a Bearer token if appropriate", async () => {
     const tokens = mockBearerTokens();
     setupOidcClientMock(tokens);
+    setupGetWebidMock("https://my.webid/", "some client ID");
     const clientCredentialsOidcHandler = new ClientCredentialsOidcHandler(
       mockDefaultTokenRefresher(),
     );
@@ -376,6 +381,7 @@ describe("handle", () => {
     const tokens = mockDpopTokens();
     tokens.refresh_token = "some refresh token";
     setupOidcClientMock(tokens);
+    setupGetWebidMock("https://my.webid/", "some client ID");
     const coreModule = jest.requireMock(
       "@inrupt/solid-client-authn-core",
     ) as typeof SolidClientAuthnCore;
@@ -412,6 +418,7 @@ describe("handle", () => {
   it("builds a fetch authenticated including the expiration value if present", async () => {
     const tokens = mockDpopTokens();
     setupOidcClientMock(tokens);
+    setupGetWebidMock("https://my.webid/", "some client ID");
     const coreModule = jest.requireMock(
       "@inrupt/solid-client-authn-core",
     ) as typeof SolidClientAuthnCore;
@@ -461,6 +468,32 @@ describe("handle", () => {
     expect(result?.isLoggedIn).toBe(true);
     expect(result?.sessionId).toBe(standardOidcOptions.sessionId);
     expect(result?.webId).toBe("https://my.webid/");
+    expect(result?.expirationDate).toBe(
+      Date.now() + DEFAULT_EXPIRATION_TIME_SECONDS * 1000,
+    );
+  });
+
+  it("returns session info, including clientAppId, with the built fetch", async () => {
+    const tokens = mockDpopTokens();
+    setupOidcClientMock(tokens);
+    setupGetWebidMock("https://my.webid/", "some client ID");
+    const clientCredentialsOidcHandler = new ClientCredentialsOidcHandler(
+      mockDefaultTokenRefresher(),
+    );
+    const result = await clientCredentialsOidcHandler.handle({
+      ...standardOidcOptions,
+      dpop: true,
+      client: {
+        clientId: "some client ID",
+        clientSecret: "some client secret",
+        clientType: "static",
+      },
+    });
+
+    expect(result?.isLoggedIn).toBe(true);
+    expect(result?.sessionId).toBe(standardOidcOptions.sessionId);
+    expect(result?.webId).toBe("https://my.webid/");
+    expect(result?.clientAppId).toBe("some client ID");
     expect(result?.expirationDate).toBe(
       Date.now() + DEFAULT_EXPIRATION_TIME_SECONDS * 1000,
     );