+++ packages/oidc-browser/src/refresh/refreshGrant.spec.ts
@@ -43,7 +43,7 @@ jest.mock("@inrupt/solid-client-authn-core", () => {
     ...actualCoreModule,
     // This works around the network lookup to the JWKS in order to validate the ID token.
     getWebidFromTokenPayload: jest.fn(() =>
-      Promise.resolve("https://my.webid/"),
+      Promise.resolve({ webId: "https://my.webid/", clientId: "some client" }),
     ),
   };
 });